Better pick handling and keyboard shortcuts
1. Improved how picks are stored in the inventory and how they are
automatically equipped after the player's current pick breaks. When a
pick breaks, the pick chosen from the inventory to be auto equipped is
the one with the highest level and with the lowest durability of all
picks of that level. The idea is to get rid of  broken picks first
before equipping and using an unused pick.

2. Ensured the "Craft" and "Craft All" buttons, and the crafting amount
textbox are disabled while the player is crafting.

3. If the player is crafting 'n' items, the crafting amount textbox will
count down from that number to let the player now how many are left to
craft.

4. If the player clicked "Craft All", the crafting amount textbox will
count up from 0 until the player runs out of resources to craft another
of the desired item. There's no indication of the total possible amount
that can be crafted, just an incrementing counter as each item is
crafted.

4. Add keyboard shortcuts. 'g' gathers resources, spacebar crafts or
cancels if already crafting, 'a' crafts all of the selected recipe, and
1-9 crafts that many of the selected recipe.

5. Made the inventory table look a little better.