Adding forges and improved recipe selection logic
1. Added all the forge items. They are not yet hooked up to handle the
smelting of ingots.

2. Added the following to the recipe details: which recipes are unlocked
by crafting it, description text, how much each item sells for, and
which, if any, forge it requires.

3. Updated the logic that selects a recipe when you click on a
requirement. Before I was tracking the corresponding page element and
calling "click()" on it. This was bad and is now replaced done properly
with KO observables only. I am still keeping track of the element for
now since jQuery requires it to cancel the crafting animation on it.

4. Added an equipment "hotbar". Slot 1 is for your pick. Slot 2 is for
your forge. Haven't decided what the remaining slots will be used for.
Perhaps some armor or trinkets that give some bonuses.

5. If the amount of an item in the inventory ever dropped to 0, I was
deleting its reference (including the observable corresponding to its
amount) from memory. This was causing a bug where, after crafting an
item, if it used up ALL of a given requirement from the inventory, it
wouldn't be removed from the inventory UI. I fixed the bug by simply
just not showing the row in the inventory table when an item's amount is
0.